backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.7
pydantic-settings>=2.2
sqlalchemy>=2.0
psycopg[binary]>=3.2
pgvector>=0.3
alembic>=1.13
python-multipart>=0.0.9
tenacity>=8.3
httpx>=0.27
openai>=1.0
google-generativeai>=0.8

# PDF extraction
pymupdf>=1.24

# GCP optional: install with pip install -r requirements-gcp.txt when using Vertex/GCS

# Utils
python-dotenv>=1.0

# Metrics
prometheus-client>=0.20
